Move coercePositiveIntegerOption into shared numeric options
Original: src/plugin/src/options/option-utils.js

New: src/shared/utils/numeric-options.js

Rationale: Numeric option coercion is shared by the printer and option normalization, so centralizing it with other numeric option helpers keeps related logic together and removes an unnecessary plugin-only wrapper.
