+++ src/shared/tests/numeric-option-utils.test.js
@@ -4,6 +4,7 @@ import test from "node:test";
 import {
     coercePositiveInteger,
     coerceNonNegativeInteger,
+    coercePositiveIntegerOption,
     resolveIntegerOption,
     normalizeNumericOption
 } from "../numeric-option-utils.js";
@@ -31,6 +32,24 @@ test("coerceNonNegativeInteger enforces a minimum of 0", () => {
                 createErrorMessage: (received) => `bad: ${received}`
             }),
         new TypeError("bad: -1")
+    );
+});
+
+test("coercePositiveIntegerOption falls back to defaults", () => {
+    assert.strictEqual(coercePositiveIntegerOption(7, 3), 7);
+    assert.strictEqual(coercePositiveIntegerOption(undefined, 3), 3);
+    assert.strictEqual(coercePositiveIntegerOption(null, 3), 3);
+    assert.strictEqual(coercePositiveIntegerOption(0, 3), 3);
+});
+
+test("coercePositiveIntegerOption respects zero replacement", () => {
+    assert.strictEqual(
+        coercePositiveIntegerOption(0, 3, { zeroReplacement: 10 }),
+        10
+    );
+    assert.strictEqual(
+        coercePositiveIntegerOption("bad", 4, { zeroReplacement: 8 }),
+        4
     );
 });
